+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 #  (2020-08-25)
 
 
+### Features
+
+- Server bootstrap ([22738a5](https://github.com/Miczeq22/brainflash-api/commit/22738a526618f41739c04fc3e6aacf9782d4576c))
+
+
 ### Documentation
 
-- README Update ([03230f2](https://github.com/Miczeq22/brainflash-api/commit/03230f25521977b58eaa6aefd93d61f95f8f817a))
+- README Update ([7c5478c](https://github.com/Miczeq22/brainflash-api/commit/7c5478c708ffbb8e63392b74cfe078f246c4a391))
 
 
 ### Miscellaneous